requirements.txt
selenium
beautifulsoup4
anthropic
openai
google-generativeai